+++ DeviceAdapters/PCO_Generic/MicroManager.cpp
@@ -6,7 +6,7 @@
 // DESCRIPTION:   pco camera module
 //                
 // AUTHOR:        Franz Reitner, Franz.Reitner@pco.de, 11/01/2010
-// COPYRIGHT:     PCO AG, Kelheim, 2010
+// COPYRIGHT:     PCO AG, Kelheim, 2010-up to now ;-)
 // LICENSE:       This file is distributed under the BSD license.
 //                License text is included with the source distribution.
 //
@@ -19,18 +19,13 @@
 //                INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES.
 //
 
-
-/*#ifdef WIN32
-#define WIN32_LEAN_AND_MEAN
-#include <windows.h>
-#endif*/
-
-#define PCO_ERRT_H_CREATE_OBJECT
-
 #include "..\..\MMDevice/ModuleInterface.h"
+#define PCO_ERRT_H_CREATE_OBJECT
+#include "PCO_err.h"
+#include "PCO_errt.h"
+
 #include "MicroManager.h"
 #include "VersionNo.h"
-//#pragma warning(disable : 4996) // disable warning for deperecated CRT functions on Windows 
 
 #if defined _WIN64
 #pragma comment(lib, ".\\lib\\pco_generic\\pco_kamlib64.lib")
@@ -57,8 +52,6 @@ const char* g_TimeStamp_BA = "Binary + ASCII";
 
 int g_iCameraCount = 0;
 int g_iSC2Count = 0;
-int g_iSenCount = 0;
-int g_iPFCount = 0;
 
 /* DEVICE_INTERFACE_VERSION is defined in MMDevice.h under MMDevice folder
 */
@@ -69,7 +62,7 @@ int g_iPFCount = 0;
 MODULE_API void InitializeModuleData()
 {
   //AddAvailableDeviceName(g_CameraDeviceName, "PCO generic camera adapter");
-  RegisterDevice(g_CameraDeviceName, MM::CameraDevice, "PCO generic camera adapter");
+  RegisterDevice(g_CameraDeviceName, MM::CameraDevice, "pco.sdk camera adapter");
 }
 
 MODULE_API void DeleteDevice(MM::Device* pDevice)
@@ -190,8 +183,6 @@ pictime_(0.0)
   m_dwFlimFrequency = 0;
   m_dwFlimPhaseMilliDeg = 0;
 
-  m_iXRes = 0;
-  m_iYRes = 0;
   m_iNextBuffer = -1;
   for(int i = 0; i < 4; i++)
   {
@@ -199,6 +190,9 @@ pictime_(0.0)
     m_iLastBufferUsed[i] = -1;
   }
   m_iNumImagesInserted = -1;
+  m_bDotPhotonFilterAvailable = false;
+  m_bDoDotPhoton = false;
+  InitDotPhoton();
 }
 
 CPCOCam::~CPCOCam()
@@ -206,13 +200,14 @@ CPCOCam::~CPCOCam()
   if(m_bInitialized)
   {
     g_iCameraCount--;
-    if(m_iInterFace == 1)
-      g_iSenCount--;
-    if(m_iInterFace == 2)
-      g_iPFCount--;
     if(m_iInterFace == 3)
       g_iSC2Count--;
     Shutdown();
+
+    if (m_bDotPhotonFilterAvailable)
+    {
+      UnInitDotPhoton();
+    }
   }
   m_bInitialized = false;
   delete(m_pCamera);
@@ -222,11 +217,9 @@ CPCOCam::~CPCOCam()
   {
     InitLib(255, NULL, 0, NULL); // Removes allocated objects
     g_iCameraCount = 0;
-    g_iSenCount = 0;
-    g_iPFCount = 0;
     g_iSC2Count = 0;
   }
-
+  
   delete(sthd_);
   CloseHandle(mxMutex);
 }
@@ -918,646 +911,6 @@ int CPCOCam::OnBinning(MM::PropertyBase* pProp, MM::ActionType eAct)
   return DEVICE_OK;
 }
 
-int CPCOCam::OnEMLeftROI(MM::PropertyBase* pProp, MM::ActionType eAct)
-{
-  int nErr = 0;
-
-  if(m_bDemoMode)
-    return DEVICE_OK;
-  nErr = 0;
-  if(eAct == MM::AfterSet)
-  {
-    long left;
-    pProp->Get(left);
-    if(m_pCamera->m_iCamClass == 2)
-    {
-      left -= 1;
-      if(left < 0)
-        left = 0;
-    }
-
-    m_nRoiXMin = left;
-    m_nRoiYMin = 1;
-    m_nRoiXMax = roiXMaxFull_;
-    m_nRoiYMax = roiYMaxFull_;
-    nErr = SetupCamera(true, true);
-    if(nErr != 0)
-    {
-      return nErr;
-    }
-    return ResizeImageBuffer();
-  }
-  else if(eAct == MM::BeforeGet)
-  {
-    if(m_bSettingsChanged)
-    {
-      /*nErr = m_pCamera->getsettings(&m_nMode, &m_nTrig, &m_nRoiXMin, &m_nRoiXMax, &m_nRoiYMin, &m_nRoiYMax,
-        &m_nHBin, &m_nVBin, m_pszTimes, &m_iGain, &m_iOffset, &m_uiFlags);
-      if(m_pCamera->m_iCamClass == 3)// pco.camera*/
-      {
-        m_pCamera->GetCameraStruct((PCO_Camera*) &m_pCamera->m_strCamera.wSize);
-      }
-      m_bSettingsChanged = FALSE;
-    }
-    if(nErr != 0)
-      return nErr;
-
-    pProp->Set((long) m_nRoiXMin);
-  }
-  return DEVICE_OK;
-}
-
-
-int CPCOCam::OnCmosParameter(MM::PropertyBase* pProp, MM::ActionType eAct)
-{
-  if(eAct == MM::BeforeGet)
-  {
-    if(m_wCMOSParameter == 0)
-      pProp->Set("Off");
-    else
-      pProp->Set("On");
-  }
-  else if(eAct == MM::AfterSet)
-  {
-    int nErr = 0;
-    string tmp;
-    long lparameter;
-    pProp->Get(tmp);
-    ((MM::Property *) pProp)->GetData(tmp.c_str(), lparameter);
-
-    if(lparameter != m_wCMOSParameter)
-    {
-      if(m_iPixelRate == 0)
-        m_wCMOSParameter = (WORD) lparameter;
-      if((m_iPixelRate == 1) && (m_pCamera->m_strCamera.strTiming.dwCMOSFlags & 0x02))
-        m_wCMOSParameter = (WORD) lparameter;
-      m_pCamera->m_strCamera.strTiming.wCMOSParameter = m_wCMOSParameter;
-      nErr = SetupCamera(true, false);
-    }
-
-    if(nErr != 0)
-      return nErr;
-  }
-  return DEVICE_OK;
-}
-
-int CPCOCam::OnCmosTimeBase(MM::PropertyBase* pProp, MM::ActionType eAct)
-{
-  if(eAct == MM::BeforeGet)
-  {
-    if(m_wCMOSTimeBase == 0)
-      pProp->Set("ns");
-    else
-    if(m_wCMOSTimeBase == 1)
-      pProp->Set("us");
-    else
-      pProp->Set("ms");
-  }
-  else if(eAct == MM::AfterSet)
-  {
-    int nErr = 0;
-    string tmp;
-    long lparameter;
-    pProp->Get(tmp);
-    ((MM::Property *) pProp)->GetData(tmp.c_str(), lparameter);
-
-    if(lparameter != m_wCMOSTimeBase)
-    {
-      if(lparameter == 0)
-      {
-        if(m_wCMOSTimeBase >= 1)
-          m_dwCMOSLineTime *= 1000;
-        if(m_wCMOSTimeBase == 2)
-          m_dwCMOSLineTime *= 1000;
-      }
-      if(lparameter == 1)
-      {
-        if(m_wCMOSTimeBase == 0)
-          m_dwCMOSLineTime /= 1000;
-        if(m_wCMOSTimeBase == 2)
-          m_dwCMOSLineTime *= 1000;
-      }
-      if(lparameter == 2)
-      {
-        if(m_wCMOSTimeBase <= 1)
-          m_dwCMOSLineTime /= 1000;
-        if(m_wCMOSTimeBase == 0)
-          m_dwCMOSLineTime /= 1000;
-      }
-      m_wCMOSTimeBase = (WORD) lparameter;
-      m_pCamera->m_strCamera.strTiming.wCMOSTimeBase = m_wCMOSTimeBase;
-
-      CheckLineTime(&m_dwCMOSLineTime);
-      m_pCamera->m_strCamera.strTiming.dwCMOSLineTime = m_dwCMOSLineTime;
-      nErr = SetupCamera(true, false);
-    }
-
-    if(nErr != 0)
-      return nErr;
-  }
-  return DEVICE_OK;
-}
-
-int CPCOCam::CheckLineTime(DWORD *ptime)
-{
-  double dmintime = m_dwCMOSLineTimeMin;
-  double dmaxtime = m_dwCMOSLineTimeMax;
-  double dtime = *ptime;
-  bool bchanged = false;
-  if(m_pCamera->m_strCamera.strTiming.wCMOSTimeBase == 0)
-  {
-    dmaxtime *= 1000000.0;
-  }
-  if(m_pCamera->m_strCamera.strTiming.wCMOSTimeBase == 1)
-  {
-    dmaxtime *= 1000.0;
-    dmintime /= 1000.0;
-  }
-  if(m_pCamera->m_strCamera.strTiming.wCMOSTimeBase == 2)
-  {
-    dmintime /= 1000000.0;
-  }
-  if(dtime > dmaxtime)
-  {
-    dtime = dmaxtime;
-    bchanged = true;
-  }
-  if(dtime < dmintime)
-  {
-    dtime = dmintime;
-    bchanged = true;
-  }
-  if(dtime < 1.0)
-    dtime = 1.0;
-
-  *ptime = (DWORD) dtime;
-  if(bchanged)
-    return DEVICE_ERR;
-  return DEVICE_OK;
-}
-
-int CPCOCam::OnCmosLineTime(MM::PropertyBase* pProp, MM::ActionType eAct)
-{
-  if(eAct == MM::BeforeGet)
-  {
-    pProp->Set((long) m_dwCMOSLineTime);
-  }
-  else if(eAct == MM::AfterSet)
-  {
-    int nErr = 0;
-    long lhelp;
-
-    ((MM::IntegerProperty*) pProp)->Get(lhelp);
-    if(lhelp != (long) m_dwCMOSLineTime)
-    {
-      m_dwCMOSLineTime = (DWORD) lhelp;
-      CheckLineTime(&m_dwCMOSLineTime);
-      m_pCamera->m_strCamera.strTiming.dwCMOSLineTime = m_dwCMOSLineTime;
-      nErr = SetupCamera(true, false);
-    }
-
-    if(nErr != 0)
-      return nErr;
-  }
-  return DEVICE_OK;
-}
-
-int CPCOCam::OnCmosExposureLines(MM::PropertyBase* pProp, MM::ActionType eAct)
-{
-  if(eAct == MM::BeforeGet)
-  {
-    pProp->Set((long) m_dwCMOSExposureLines);
-  }
-  else if(eAct == MM::AfterSet)
-  {
-    int nErr = 0;
-    long lhelp;
-
-    ((MM::IntegerProperty*) pProp)->Get(lhelp);
-    if(lhelp != (long) m_dwCMOSExposureLines)
-    {
-      if(lhelp < 1)
-        lhelp = 1;
-      if(lhelp > 0x7FFFFFFF)
-        lhelp = (long) 0x7FFFFFFF;
-      m_dwCMOSExposureLines = lhelp;
-      m_pCamera->m_strCamera.strTiming.dwCMOSExposureLines = m_dwCMOSExposureLines;
-      nErr = SetupCamera(false, false);
-    }
-
-    if(nErr != 0)
-      return nErr;
-  }
-  return DEVICE_OK;
-}
-
-int CPCOCam::OnCmosDelayLines(MM::PropertyBase* pProp, MM::ActionType eAct)
-{
-  if(eAct == MM::BeforeGet)
-  {
-    pProp->Set((long) m_dwCMOSDelayLines);
-  }
-  else if(eAct == MM::AfterSet)
-  {
-    int nErr = 0;
-    long lhelp;
-
-    ((MM::IntegerProperty*) pProp)->Get(lhelp);
-    if(lhelp != (long) m_dwCMOSDelayLines)
-    {
-      if(lhelp < 0)
-        lhelp = 0;
-      if(lhelp > 0x7FFFFFFF)
-        lhelp = (long) 0x7FFFFFFF;
-      m_dwCMOSDelayLines = lhelp;
-      m_pCamera->m_strCamera.strTiming.dwCMOSDelayLines = m_dwCMOSDelayLines;
-      nErr = SetupCamera(false, false);
-    }
-
-    if(nErr != 0)
-      return nErr;
-  }
-  return DEVICE_OK;
-}
-
-int CPCOCam::OnFlimModulationSource(MM::PropertyBase* pProp, MM::ActionType eAct)
-{
-  if(eAct == MM::BeforeGet)
-  {
-    if(m_wFlimSourceSelect == 0)
-      pProp->Set("(0) intern");
-    else
-      pProp->Set("(1) extern");
-  }
-  else if(eAct == MM::AfterSet)
-  {
-    int nErr = 0;
-    string tmp;
-    long lparameter;
-    pProp->Get(tmp);
-    ((MM::Property *) pProp)->GetData(tmp.c_str(), lparameter);
-
-    if(lparameter != m_wFlimSourceSelect)
-    {
-      m_wFlimSourceSelect = (WORD) lparameter;
-
-      nErr = SetupFlim();
-      if(nErr != 0) // A return error refuses to set to extern, so reset value and update settings
-      {
-        m_wFlimSourceSelect = 0;
-        UpdateStatus();
-      }
-    }
-
-    if(nErr != 0)
-      return nErr;
-  }
-  return DEVICE_OK;
-}
-
-int CPCOCam::OnFlimMasterFrequencyMHz(MM::PropertyBase* pProp, MM::ActionType eAct)
-{
-  if(eAct == MM::BeforeGet)
-  {
-    if(m_bFlimMasterFrequencyMHz)
-      pProp->Set("(1) Mhz");
-    else
-      pProp->Set("(0) kHz");
-  }
-  else if(eAct == MM::AfterSet)
-  {
-    string tmp;
-    long frequTmp;
-    pProp->Get(tmp);
-    ((MM::Property *) pProp)->GetData(tmp.c_str(), frequTmp);
-    if(m_bFlimMasterFrequencyMHz != (frequTmp == 1))
-    {
-      if(m_wFlimSourceSelect == 0) // Changes are transferred only when source is set to intern
-      {
-        m_bFlimMasterFrequencyMHz = (frequTmp == 1);
-        SetupFlim();
-      }
-      else
-        UpdateStatus();
-    }
-  }
-  return DEVICE_OK;
-}
-
-
-int CPCOCam::OnFlimFrequency(MM::PropertyBase* pProp, MM::ActionType eAct)
-{
-  long lhelp;
-  if(eAct == MM::BeforeGet)
-  {
-    lhelp = m_dwFlimFrequency;
-    if(m_bFlimMasterFrequencyMHz)
-      lhelp /= 1000;
-    pProp->Set(lhelp);
-  }
-  else if(eAct == MM::AfterSet)
-  {
-    int nErr = 0;
-
-    ((MM::IntegerProperty*) pProp)->Get(lhelp);
-    if(lhelp != (long) m_dwFlimFrequency)
-    {
-      if(m_wFlimSourceSelect == 0) // Changes are transferred only when source is set to intern
-      {
-        if(m_bFlimMasterFrequencyMHz)
-          lhelp *= 1000;
-
-        if(lhelp < 0)
-          lhelp = 0;
-        if(lhelp > 50000000)
-          lhelp = 50000000;
-        m_dwFlimFrequency = lhelp;
-        nErr = SetupFlim();
-      }
-      else
-        UpdateStatus();
-    }
-
-    if(nErr != 0)
-      return nErr;
-  }
-  return DEVICE_OK;
-}
-
-int CPCOCam::OnFlimRelativePhase(MM::PropertyBase* pProp, MM::ActionType eAct)
-{
-  if(eAct == MM::BeforeGet)
-  {
-    pProp->Set((long) m_dwFlimPhaseMilliDeg);
-  }
-  else if(eAct == MM::AfterSet)
-  {
-    int nErr = 0;
-    long lhelp;
-
-    ((MM::IntegerProperty*) pProp)->Get(lhelp);
-    if(lhelp != (long) m_dwFlimPhaseMilliDeg)
-    {
-      if(m_wFlimPhaseNumber == 0)// Must be 0 and not 2,4,8 or 16 (!=0)
-      {
-        if(lhelp < 0)
-          lhelp = 0;
-        if(lhelp > 359999)
-          lhelp = 359999;
-
-        m_dwFlimPhaseMilliDeg = lhelp;
-        nErr = SetupFlim();
-      }
-      else
-        UpdateStatus();
-    }
-
-    if(nErr != 0)
-      return nErr;
-  }
-  return DEVICE_OK;
-}
-
-int CPCOCam::OnFlimOutputWaveForm(MM::PropertyBase* pProp, MM::ActionType eAct)
-{
-  if(eAct == MM::BeforeGet)
-  {
-    if(m_wFlimOutputWaveform == 0)
-      pProp->Set("(0) none");
-    else
-    if(m_wFlimOutputWaveform == 1)
-      pProp->Set("(1) sine wave");
-    else
-      pProp->Set("(2) square wave");
-  }
-  else if(eAct == MM::AfterSet)
-  {
-    int nErr = 0;
-    string tmp;
-    long waveTmp;
-    pProp->Get(tmp);
-    ((MM::Property *) pProp)->GetData(tmp.c_str(), waveTmp);
-
-    if(waveTmp != m_wFlimOutputWaveform)
-    {
-      if(waveTmp < 0)
-        waveTmp = 0;
-      if(waveTmp > 2)
-        waveTmp = 2;
-      m_wFlimOutputWaveform = (WORD) waveTmp;
-      nErr = SetupFlim();
-    }
-
-    if(nErr != 0)
-      return nErr;
-  }
-  return DEVICE_OK;
-}
-
-int CPCOCam::OnFlimNumberOfPhaseSamples(MM::PropertyBase* pProp, MM::ActionType eAct)
-{
-  if(eAct == MM::BeforeGet)
-  {
-    switch(m_wFlimPhaseNumber)
-    {
-      case 0:
-        pProp->Set("(0) shiftable pair");
-      break;
-      case 1:
-        pProp->Set("(1) 2");
-      break;
-      case 2:
-        pProp->Set("(2) 4");
-      break;
-      case 3:
-        pProp->Set("(3) 8");
-      break;
-      case 4:
-        pProp->Set("(4) 16");
-      break;
-    }
-  }
-  else if(eAct == MM::AfterSet)
-  {
-    int nErr = 0;
-    string tmp;
-    long waveTmp;
-    pProp->Get(tmp);
-    ((MM::Property *) pProp)->GetData(tmp.c_str(), waveTmp);
-
-    if(waveTmp != m_wFlimPhaseNumber)
-    {
-      if(waveTmp < 0)
-        waveTmp = 0;
-      if(waveTmp > 4)
-        waveTmp = 4;
-
-      m_wFlimPhaseNumber = (WORD) waveTmp;
-      nErr = SetupFlim();
-    }
-
-    if(nErr != 0)
-      return nErr;
-  }
-  return DEVICE_OK;
-}
-
-int CPCOCam::OnFlimPhaseSymmetry(MM::PropertyBase* pProp, MM::ActionType eAct)
-{
-  if(eAct == MM::BeforeGet)
-  {
-    if(m_wFlimPhaseSymmetry == 0)
-      pProp->Set("(0) no");
-    else
-      pProp->Set("(1) yes");
-  }
-  else if(eAct == MM::AfterSet)
-  {
-    int nErr = 0;
-    string tmp;
-    long waveTmp;
-    pProp->Get(tmp);
-    ((MM::Property *) pProp)->GetData(tmp.c_str(), waveTmp);
-
-    if(waveTmp != m_wFlimPhaseSymmetry)
-    {
-      if(m_wFlimPhaseNumber != 0)
-      {
-        if(waveTmp < 0)
-          waveTmp = 0;
-        if(waveTmp > 1)
-          waveTmp = 1;
-
-        m_wFlimPhaseSymmetry = (WORD) waveTmp;
-        nErr = SetupFlim();
-      }
-      else
-        UpdateStatus();
-    }
-
-    if(nErr != 0)
-      return nErr;
-  }
-  return DEVICE_OK;
-}
-
-int CPCOCam::OnFlimPhaseOrder(MM::PropertyBase* pProp, MM::ActionType eAct)
-{
-  if(eAct == MM::BeforeGet)
-  {
-    if(m_wFlimPhaseOrder == 0)
-      pProp->Set("(0) ascending");
-    else
-      pProp->Set("(1) opposite");
-  }
-  else if(eAct == MM::AfterSet)
-  {
-    int nErr = 0;
-    string tmp;
-    long waveTmp;
-    pProp->Get(tmp);
-    ((MM::Property *) pProp)->GetData(tmp.c_str(), waveTmp);
-
-    if(waveTmp != m_wFlimPhaseOrder)
-    {
-      if(m_wFlimPhaseSymmetry == 1)  // Must be Yes (1)
-      {
-        if(waveTmp < 0)
-          waveTmp = 0;
-        if(waveTmp > 1)
-          waveTmp = 1;
-
-        m_wFlimPhaseOrder = (WORD) waveTmp;
-        nErr = SetupFlim();
-      }
-      else
-      {
-        m_wFlimPhaseOrder = 0;
-        UpdateStatus();
-      }
-    }
-
-    if(nErr != 0)
-      return nErr;
-  }
-  return DEVICE_OK;
-}
-
-int CPCOCam::OnFlimTapSelection(MM::PropertyBase* pProp, MM::ActionType eAct)
-{
-  if(eAct == MM::BeforeGet)
-  {
-    switch(m_wFlimTapSelect)
-    {
-      case 0:
-        pProp->Set("(0) Tap A + B");
-      break;
-      case 1:
-        pProp->Set("(1) Tap A");
-      break;
-      case 2:
-        pProp->Set("(2) Tap B");
-      break;
-    }
-  }
-  else if(eAct == MM::AfterSet)
-  {
-    int nErr = 0;
-    string tmp;
-    long waveTmp;
-    pProp->Get(tmp);
-    ((MM::Property *) pProp)->GetData(tmp.c_str(), waveTmp);
-
-    if(waveTmp != m_wFlimTapSelect)
-    {
-      if(waveTmp < 0)
-        waveTmp = 0;
-      if(waveTmp > 2)
-        waveTmp = 2;
-
-      m_wFlimTapSelect = (WORD) waveTmp;
-      nErr = SetupFlim();
-    }
-
-    if(nErr != 0)
-      return nErr;
-  }
-  return DEVICE_OK;
-}
-
-int CPCOCam::OnFlimAsymCorrection(MM::PropertyBase* pProp, MM::ActionType eAct)
-{
-  if(eAct == MM::BeforeGet)
-  {
-    if(m_wFlimAsymmetryCorrection == 0)
-      pProp->Set("(0) off");
-    else
-      pProp->Set("(1) on");
-  }
-  else if(eAct == MM::AfterSet)
-  {
-    int nErr = 0;
-    string tmp;
-    long waveTmp;
-    pProp->Get(tmp);
-    ((MM::Property *) pProp)->GetData(tmp.c_str(), waveTmp);
-
-    if(waveTmp != m_wFlimAsymmetryCorrection)
-    {
-      if(waveTmp < 0)
-        waveTmp = 0;
-      if(waveTmp > 1)
-        waveTmp = 1;
-
-      m_wFlimAsymmetryCorrection = (WORD) waveTmp;
-      nErr = SetupFlim();
-    }
-
-    if(nErr != 0)
-      return nErr;
-  }
-  return DEVICE_OK;
-}
 
 
 int CPCOCam::SetupCamera(bool bStopRecording, bool bSizeChanged)
@@ -1596,26 +949,12 @@ int CPCOCam::SetupCamera(bool bStopRecording, bool bSizeChanged)
   m_nTimesLen = MM_PCO_GENERIC_MAX_STRLEN;
   if((m_nCameraType == 0x1300) || (m_nCameraType == 0x1310))
     iOffsPxr = m_iPixelRate;
-  //if(m_pCamera->m_iCamClass == 3)
-  {
-    DWORD dwresult = 0;
-    m_pCamera->SetCameraStruct((PCO_Camera*) &m_pCamera->m_strCamera.wSize, &dwresult);
 
-    std:string szsetname = "";
-    m_pCamera->WriteSettingsToRegistryMM(0, szsetname, m_pCamera->m_strCamera);
-  }
-  /*else
-  {
-    nErr = m_pCamera->testcoc(&m_nMode, &m_nTrig, &m_nRoiXMin, &m_nRoiXMax, &m_nRoiYMin, &m_nRoiYMax,
-    &m_nHBin, &m_nVBin, m_pszTimes, &m_nTimesLen, &m_iGain, &iOffsPxr, &m_uiFlags);
-    if ((nErr != 0) && (nErr != 103))
-      return nErr;
+  DWORD dwresult = 0;
+  m_pCamera->SetCameraStruct((PCO_Camera*) &m_pCamera->m_strCamera.wSize, &dwresult);
 
-    nErr = m_pCamera->setcoc(m_nMode, m_nTrig, m_nRoiXMin, m_nRoiXMax, m_nRoiYMin, m_nRoiYMax,
-    m_nHBin, m_nVBin, m_pszTimes, m_iGain, iOffsPxr, m_uiFlags);
-    if (nErr != 0)
-      return nErr;
-  }*/
+  char szsetname[100] = "";
+  m_pCamera->WriteSettingsToRegistryMM(0, &szsetname[0], m_pCamera->m_strCamera);
 
   m_bSettingsChanged = TRUE;
   if(bStopRecording || m_bSoftwareTriggered)
@@ -1691,61 +1030,41 @@ int CPCOCam::OnPixelType(MM::PropertyBase* pProp, MM::ActionType eAct)
 
 int CPCOCam::OnGain(MM::PropertyBase* pProp, MM::ActionType eAct)
 {
-  if(m_pCamera->m_iCamClass == 1)
+  if(m_pCamera->m_iCamClass == 3)
   {
     if(eAct == MM::BeforeGet)
     {
-      if(m_iGainCam == 0)
-        pProp->Set("normal");
-      else
-      if(m_iGainCam == 1)
-        pProp->Set("extended");
-      else
-      if(m_iGainCam == 3)
-        pProp->Set("low light mode");
+      char szval[20] = { 0 };
+      snprintf(szval, 20, "%d", (int)m_pCamera->m_strCamera.strSensor.wConvFact);
+      m_iGainCam = m_pCamera->m_strCamera.strSensor.wConvFact;
+
+      pProp->Set(szval);
     }
     else if(eAct == MM::AfterSet)
     {
-      int igains[3] = {0, 1, 3};
+      int igains[4] = {0, 1, 2, 3};
       int nErr = 0;
       long ihelp;
       string tmp;
       pProp->Get(tmp);
       ((MM::Property *) pProp)->GetData(tmp.c_str(), ihelp);
-      ihelp = igains[ihelp];
+
       if(ihelp != m_iGainCam)
       {
         m_iGainCam = ihelp;
-        ihelp = m_nMode & 0xFF;
-        m_nMode = ihelp + (m_nSubMode << 16) + (m_iGainCam << 8);
+
+        m_pCamera->m_strCamera.strSensor.wConvFact = (WORD)m_iGainCam;
+
         nErr = SetupCamera(true, false);
+        if (nErr != 0)
+        {
+          return nErr;
+        }
       }
 
       if(nErr != 0)
         return nErr;
     }
-  }
-  return DEVICE_OK;
-}
-
-int CPCOCam::OnEMGain(MM::PropertyBase* pProp, MM::ActionType eAct)
-{
-  if(eAct == MM::BeforeGet)
-    pProp->Set((long) m_iEMGain);
-  else if(eAct == MM::AfterSet)
-  {
-    int nErr = 0;
-    long ihelp;
-    pProp->Get(ihelp);
-    if(ihelp != m_iEMGain)
-    {
-      m_iEMGain = ihelp;
-      sprintf_s(m_pszTimes, sizeof(m_pszTimes), "0,%d,-1,-1\r\nmg%d", (int) m_dExposure, m_iEMGain);
-      nErr = SetupCamera(true, false);
-    }
-
-    if(nErr != 0)
-      return nErr;
   }
   return DEVICE_OK;
 }
@@ -1802,12 +1121,12 @@ int CPCOCam::Initialize()
     return nRet;
 
   // Description
-  nRet = CreateProperty("Description", "pco generic driver module", MM::String, true);
+  nRet = CreateProperty("Description", "pco.sdk driver module", MM::String, true);
   if(nRet != DEVICE_OK)
     return nRet;
 
   char szVersion[200];
-  sprintf_s(szVersion, 200, "pco generic device adapter - V%s", STRFILEVER);
+  sprintf_s(szVersion, 200, "pco.sdk device adapter - V%s", STRFILEVER);
   LogMessage(szVersion, false);
 
   CPropertyAction* pAct;
@@ -1822,32 +1141,7 @@ int CPCOCam::Initialize()
     nErr = m_pCamera->PreInitSC2(g_iSC2Count, g_iCameraCount, &uiresult);
     if(nErr != PCO_NOERROR)
     {
-      EnableConvert(TRUE);
-      nErr = m_pCamera->PreInitSen(g_iSenCount, g_iCameraCount, &uiresult);
-      if(nErr != PCO_NOERROR)
-      {
-        nErr = m_pCamera->PreInitPcCam(g_iPFCount, g_iCameraCount, &uiresult);
-        if(nErr != PCO_NOERROR)
-        {
-          return DEVICE_NOT_CONNECTED;
-        }
-        else
-        {
-          m_iCameraNum = g_iCameraCount;
-          m_iCameraNumAtInterface = g_iPFCount;
-          m_iInterFace = 2;
-          g_iPFCount++;
-          g_iCameraCount++;
-        }
-      }
-      else
-      {
-        m_iCameraNum = g_iCameraCount;
-        m_iCameraNumAtInterface = g_iSenCount;
-        m_iInterFace = 1;
-        g_iSenCount++;
-        g_iCameraCount++;
-      }
+      return DEVICE_NOT_CONNECTED;
     }
     else
     {
@@ -1888,6 +1182,7 @@ int CPCOCam::Initialize()
     WriteLog("pco_generic. Error %x in Init!", nErr);
     return nErr;
   }
+
   // camera type (read-only)
   pAct = new CPropertyAction(this, &CPCOCam::OnCameraType);
   nRet = CreateProperty("CameraType", "", MM::String, true, pAct);
@@ -1900,6 +1195,17 @@ int CPCOCam::Initialize()
   m_nCameraType = m_pCamera->GetCamType();
   m_nCCDType = m_pCamera->GetCCDType();
 
+
+  if (m_bDotPhotonFilterAvailable)
+  {
+    int err = InitDotPhotonCamera(m_pCamera->GetCameraHandle());
+    if (err != PCO_NOERROR)
+    {
+      m_bDotPhotonFilterAvailable = false;
+      m_bDoDotPhoton = false;
+    }
+  }
+
   char tab[400];
   m_pCamera->getsettings(&m_nMode, &m_nTrig, &m_nRoiXMin, &m_nRoiXMax, &m_nRoiYMin, &m_nRoiYMax,
     &m_nHBin, &m_nVBin, tab, &m_iGain, &m_iOffset, &m_uiFlags);
@@ -1909,10 +1215,11 @@ int CPCOCam::Initialize()
   if(m_pCamera->m_iCamClass == 3)
   {
     int tb[3] = {1000000, 1000, 1};
-    std:string szsetname = "";
+    char szsetname[100] = "";
+    int i = 0;
 
     m_pCamera->m_strCamera.wSize = sizeof(PCO_Camera);
-    m_pCamera->LoadSettingsFromRegistryMM(0, szsetname, m_pCamera->m_strCamera);
+    m_pCamera->LoadSettingsFromRegistryMM(i, &szsetname[0], 100, m_pCamera->m_strCamera);
     m_pCamera->GetCameraStruct((PCO_Camera*) &m_pCamera->m_strCamera.wSize);
     //m_nHBin = m_nVBin = 1;
     //m_nMode = 0;//M_LONG;
@@ -1960,76 +1267,10 @@ int CPCOCam::Initialize()
   }
   else
   {
-    ::MessageBox(NULL, "As of 23rd of August 2020 pco ag has stopped software support for SensiCam and PixelFly (PCO PCI 540).\r\nPlease use a previous version of MicroManager.\r\nAdapter still works, but the next version will not.\r\nTo avoid this dialog downgrade MicroManager.", "Deprecated camera!", MB_ICONERROR);
+    return DEVICE_NOT_CONNECTED;
   }
 
-  if(m_pCamera->m_iCamClass == 2)
-  {
-    m_nHBin = m_nVBin = 0;
-    m_nMode = 0x20000;//M_LONG;
-    m_nSubMode = 0;//NORMALLONG;
-    m_nTrig = 1;
-    sprintf_s(m_pszTimes, sizeof(m_pszTimes), "%f", m_dExposure / 1000.0);
-  }
-  if(m_pCamera->m_iCamClass == 1)
-  {
-    m_nRoiXMin = m_pCamera->m_strCamera.strSensor.wRoiX0;
-    m_nRoiYMin = m_pCamera->m_strCamera.strSensor.wRoiY0;
-    m_nRoiXMax = m_pCamera->m_strCamera.strSensor.wRoiX1;
-    m_nRoiYMax = m_pCamera->m_strCamera.strSensor.wRoiY1;
 
-    m_nHBin = m_nVBin = 1;
-    switch(m_nCameraType)
-    {
-      case 1: // Fast Shutter
-        m_nMode = 1;//M_FAST;
-        m_nSubMode = 0;//NORMALFAST;
-        m_iGainCam = 0;
-        m_nTrig = 0;
-        //1ms exposure time
-        m_dExposure = 1;
-        sprintf_s(m_pszTimes, sizeof(m_pszTimes), "0,%.0f,-1,-1", m_dExposure * 1.0e06);
-      break;
-
-      case 2:
-      case 7:
-        // Long exposure 2; Long exposure QE 7
-        m_nMode = 0;//M_LONG;
-        m_nSubMode = 0;//NORMALLONG;
-        m_iGainCam = 0;
-        m_nTrig = 0;
-        //10ms exposure time
-
-        if(m_nCCDType == 0x21)
-          sprintf_s(m_pszTimes, sizeof(m_pszTimes), "0,%.0f,-1,-1\r\nmg%d", m_dExposure, m_iEMGain);
-        else
-          sprintf_s(m_pszTimes, sizeof(m_pszTimes), "0,%.0f,-1,-1", m_dExposure);
-      break;
-
-      case 3:
-        // Long exposure OEM
-        m_nMode = 0;//M_LONG;
-        m_nSubMode = 0;//NORMALLONG;
-        m_nTrig = 0;
-        //10ms exposure time
-        sprintf_s(m_pszTimes, sizeof(m_pszTimes), "0,%.0f,-1,-1", m_dExposure);
-      break;
-
-      case 8://FASTEXPQE
-        m_nMode = 0;//M_LONG;
-        m_nSubMode = 0;//NORMALLONG;
-        m_iGainCam = 0;
-        m_nTrig = 0;
-        //10ms exposure time
-        sprintf_s(m_pszTimes, sizeof(m_pszTimes), "0,%.0f,-1,-1", m_dExposure);
-      break;
-
-      default:
-      // invalid type
-        return ERR_UNKNOWN_CAMERA_TYPE;
-      break;
-    }
-  }
   int imode, isubmode, igain;
 
   m_pCamera->ReloadSize();
@@ -2051,16 +1292,7 @@ int CPCOCam::Initialize()
 
   vector<string> binValues;
   binValues.push_back("1");
-  if(m_pCamera->m_iCamClass == 1)
-  {
-    binValues.push_back("2");
-    binValues.push_back("4");
-    binValues.push_back("8");
-  }
-  if(m_pCamera->m_iCamClass == 2)
-  {
-    binValues.push_back("2");
-  }
+
   if(m_pCamera->m_iCamClass == 3)
   {
     if((m_pCamera->m_strCamera.strSensor.strDescription.wMaxBinHorzDESC >= 2) &&
@@ -2108,76 +1340,11 @@ int CPCOCam::Initialize()
   if(nRet != DEVICE_OK)
     return nRet;
 
-  if(m_pCamera->m_iCamClass == 1)
-  {
-    // EMGain
-    if((m_nCCDType == 0x21) ||  // TI EM 285
-      (m_nCCDType == 0x27))
-    {
-      pAct = new CPropertyAction(this, &CPCOCam::OnEMGain);
-      nRet = CreateProperty(MM::g_Keyword_EMGain, "1", MM::Integer, false, pAct);
-      if(nRet != DEVICE_OK)
-        return nRet;
-      nRet = SetPropertyLimits(MM::g_Keyword_EMGain, 1.0, 9.0);
-      if(nRet != DEVICE_OK)
-        return nRet;
-
-      pAct = new CPropertyAction(this, &CPCOCam::OnEMLeftROI);
-      nRet = CreateProperty("EM left ROI", "1", MM::Integer, false, pAct);
-      if(nRet != DEVICE_OK)
-        return nRet;
-
-
-      vector<string> roiValues;
-      roiValues.push_back("1");
-      roiValues.push_back("2");
-
-      nRet = SetAllowedValues("EM left ROI", roiValues);
-      if(nRet != DEVICE_OK)
-        return nRet;
-      UpdateProperty("EM left ROI");
-
-    }
-    else
-    {
-      if(m_nCameraType == 7)
-      {
-        pAct = new CPropertyAction(this, &CPCOCam::OnGain);
-        nRet = CreateProperty("Gain", "normal", MM::String, false, pAct);
-        if(nRet != DEVICE_OK)
-          return nRet;
-        nRet = AddAllowedValue("Gain", "normal", 0);
-        if(nRet != DEVICE_OK)
-          return nRet;
-        nRet = AddAllowedValue("Gain", "extended", 1);
-        if(nRet != DEVICE_OK)
-          return nRet;
-        nRet = AddAllowedValue("Gain", "low light mode", 2);
-        if(nRet != DEVICE_OK)
-          return nRet;
-      }
-    }
-
-  }
   // Exposure
   pAct = new CPropertyAction(this, &CPCOCam::OnExposure);
   nRet = CreateProperty(MM::g_Keyword_Exposure, "10", MM::Float, false, pAct);
   if(nRet != DEVICE_OK)
     return nRet;
-
-  if(m_pCamera->m_iCamClass == 2)
-  {
-    pAct = new CPropertyAction(this, &CPCOCam::OnTriggerMode);
-    nRet = CreateProperty("Triggermode", "Internal", MM::String, false, pAct);
-    if(nRet != DEVICE_OK)
-      return nRet;
-    nRet = AddAllowedValue("Triggermode", "Internal", 1);
-    if(nRet != DEVICE_OK)
-      return nRet;
-    nRet = AddAllowedValue("Triggermode", "External", 0);
-    if(nRet != DEVICE_OK)
-      return nRet;
-  }
 
   if(m_pCamera->m_iCamClass == 3)
   {
@@ -2319,7 +1486,25 @@ int CPCOCam::Initialize()
       if(nRet != DEVICE_OK)
         return nRet;
     }
-    if(m_pCamera->m_strCamera.strSensor.strDescription.dwGeneralCapsDESC1 & GENERALCAPS1_HW_IO_SIGNAL_DESCRIPTOR)
+    if (m_pCamera->m_strCamera.strSensor.strDescription.wConvFactDESC[1] != 0)
+    {
+      char szval[20] = { 0 };
+      pAct = new CPropertyAction(this, &CPCOCam::OnGain);
+      snprintf(szval, 20, "%d", (int)m_pCamera->m_strCamera.strSensor.strDescription.wConvFactDESC[0]);
+      nRet = CreateProperty("Conversion factor*100", szval, MM::String, false, pAct);
+      if (nRet != DEVICE_OK)
+        return nRet;
+      for (int i = 0; i < 4; i++)
+      {
+        if (m_pCamera->m_strCamera.strSensor.strDescription.wConvFactDESC[i] == 0)
+          break;
+        snprintf(szval, 20, "%d", (int)m_pCamera->m_strCamera.strSensor.strDescription.wConvFactDESC[i]);
+        nRet = AddAllowedValue("Conversion factor*100", szval, m_pCamera->m_strCamera.strSensor.strDescription.wConvFactDESC[i]);
+        if (nRet != DEVICE_OK)
+          return nRet;
+      }
+    }
+    if (m_pCamera->m_strCamera.strSensor.strDescription.dwGeneralCapsDESC1 & GENERALCAPS1_HW_IO_SIGNAL_DESCRIPTOR)
       InitHWIO();
     if(m_bCMOSLineTiming)
       InitLineTiming();
@@ -2329,11 +1514,7 @@ int CPCOCam::Initialize()
     InitFanLED();
   }
 
-  //test if SET_COC gets right values
-  if(m_pCamera->m_iCamClass == 1)
-    m_nMode = imode + (isubmode << 16) + (igain << 8);
-  else
-    m_nMode = imode;
+  m_nMode = imode;
   m_nSubMode = isubmode;
   m_iGain = igain;
   nErr = SetupCamera(true, true);
@@ -2345,1064 +1526,6 @@ int CPCOCam::Initialize()
   return DEVICE_OK;
 }
 
-int CPCOCam::GetSignalNum(std::string szSigName)
-{
-  int isc = 0;
-  if(szSigName.find('1') != std::string::npos)
-    isc = 0;
-  if(szSigName.find('2') != std::string::npos)
-    isc = 1;
-  if(szSigName.find('3') != std::string::npos)
-    isc = 2;
-  if(szSigName.find('4') != std::string::npos)
-    isc = 3;
-  return isc;
-}
-
-char szSelectSignalTiming[4][40] = {"Show time of 'First Line'", "Show common time of 'All Lines'", "Show time of 'Last Line'", "Show overall time of 'All Lines'"};
-#define HWIOBUFLEN 150
-int CPCOCam::OnSelectSignal(MM::PropertyBase* pProp, MM::ActionType eAct)
-{
-  if(m_pCamera->m_iCamClass != 3)// pco.camera
-    return DEVICE_OK;
-
-  long isc = 0;//GetSignalNum(pProp->GetName());
-  string szval;
-  pProp->Get(szval);
-  MM::Property *pprophelper = (MM::Property *)pProp;
-  pprophelper->GetData(szval.c_str(), isc);
-  isc /= 0x100;
-  isc--;
-
-  if(eAct == MM::BeforeGet)
-  {
-    pProp->Set(m_pCamera->m_strCamera.strSensor.strSignalDesc.strSingeSignalDesc[isc].strSignalName[m_pCamera->m_strCamera.strTiming.strSignal[isc].wSelected]);
-  }
-  else if(eAct == MM::AfterSet)
-  {
-    std::string szselectedsignal;
-
-    pProp->Get(szselectedsignal);
-
-    string szsignal;
-    for(int i = 0; i < 4; i++)
-    {
-      szsignal = m_pCamera->m_strCamera.strSensor.strSignalDesc.strSingeSignalDesc[isc].strSignalName[i];
-      if(szsignal == szselectedsignal)
-      {
-        if(m_pCamera->m_strCamera.strTiming.strSignal[isc].wSelected != (WORD) i)
-        {
-          int isignal = i;
-          int iflag = 1 << isignal;
-
-          if(iflag & m_pCamera->m_strCamera.strSensor.strSignalDesc.strSingeSignalDesc[isc].wSignalDefinitions)
-          {
-            char csh[200];
-            int ivalue_helper = (isc + 1) * 0x100;
-            sprintf_s(csh, 200, "Signal %d (%s) Timing", isc + 1, m_pCamera->m_strCamera.strSensor.strSignalDesc.strSingeSignalDesc[isc].strSignalName[0]);
-            ClearAllowedValues(csh);
-
-            if(m_pCamera->m_strCamera.strTiming.strSignal[isc].dwSignalFunctionality[isignal] == 0x07)
-            {
-              AddAllowedValue(csh, szSelectSignalTiming[0], 0 + ivalue_helper);
-              AddAllowedValue(csh, szSelectSignalTiming[1], 1 + ivalue_helper);
-              AddAllowedValue(csh, szSelectSignalTiming[2], 2 + ivalue_helper);
-              AddAllowedValue(csh, szSelectSignalTiming[3], 3 + ivalue_helper);
-            }
-            else
-            {
-              AddAllowedValue(csh, "Not available", ivalue_helper);
-            }
-          }
-          m_pCamera->m_strCamera.strTiming.strSignal[isc].wSelected = (WORD) i;
-          SetupCamera(true, false);
-        }
-        break;
-      }
-    }
-  }
-
-  return DEVICE_OK;
-}
-
-int iBitToIndex[16] = {0, 0, 1, 1, 2, 2, 2, 2, 3, 3, 3, 3, 3, 3, 3, 3};
-int CPCOCam::OnSelectSignalTiming(MM::PropertyBase* pProp, MM::ActionType eAct)
-{
-  if(m_pCamera->m_iCamClass != 3)// pco.camera
-    return DEVICE_OK;
-
-  long isc = 0;//GetSignalNum(pProp->GetName());
-  string szval;
-  pProp->Get(szval);
-  MM::Property *pprophelper = (MM::Property *)pProp;
-  if(!pprophelper->GetData(szval.c_str(), isc))
-    return DEVICE_OK;
-
-  isc /= 0x100;
-  isc--;
-  
-  int iselectedsignal = m_pCamera->m_strCamera.strTiming.strSignal[isc].wSelected;
-  int iselectedpar = m_pCamera->m_strCamera.strTiming.strSignal[isc].dwParameter[iselectedsignal];
-  if(eAct == MM::BeforeGet)
-  {
-
-
-    if(m_pCamera->m_strCamera.strTiming.strSignal[isc].dwSignalFunctionality[iselectedsignal] == 0x07)
-      pProp->Set(szSelectSignalTiming[iselectedpar - 1]);
-    else
-      pProp->Set("Not available");
-  }
-  else if(eAct == MM::AfterSet)
-  {
-    std::string szselectedsignal;
-
-    pProp->Get(szselectedsignal);
-
-    string szsignal;
-    for(int i = 1; i <= 4; i++)
-    {
-      szsignal = szSelectSignalTiming[i - 1];
-      if(szsignal == szselectedsignal)
-      {
-        if(m_pCamera->m_strCamera.strTiming.strSignal[isc].dwParameter[iselectedsignal] != (DWORD) i)
-        {
-          m_pCamera->m_strCamera.strTiming.strSignal[isc].dwParameter[iselectedsignal] = (DWORD) i;
-          SetupCamera(true, false);
-        }
-        break;
-      }
-    }
-  }
-
-  return DEVICE_OK;
-}
-
-int CPCOCam::OnSelectSignalOnOff(MM::PropertyBase* pProp, MM::ActionType eAct)
-{
-  if(m_pCamera->m_iCamClass != 3)// pco.camera
-    return DEVICE_OK;
-
-  long isc = 0;//GetSignalNum(pProp->GetName());
-  string szval;
-  pProp->Get(szval);
-  MM::Property *pprophelper = (MM::Property *)pProp;
-  pprophelper->GetData(szval.c_str(), isc);
-  isc /= 0x100;
-  isc--;
-
-  if(eAct == MM::BeforeGet)
-  {
-    if(m_pCamera->m_strCamera.strTiming.strSignal[isc].wEnabled)
-      pProp->Set("on");
-    else
-      pProp->Set("off");
-  }
-  else if(eAct == MM::AfterSet)
-  {
-    std::string szselectedsignal;
-    int inewval = 0;
-
-    pProp->Get(szselectedsignal);
-
-    if(szselectedsignal == "on")
-      inewval = 1;
-    else
-      inewval = 0;
-    if(m_pCamera->m_strCamera.strTiming.strSignal[isc].wEnabled != (WORD) inewval)
-    {
-      m_pCamera->m_strCamera.strTiming.strSignal[isc].wEnabled = (WORD) inewval;
-      SetupCamera(true, false);
-    }
-  }
-
-  return DEVICE_OK;
-}
-
-char szSelectSignalType[5][40] = {"TTL", "High Level", "Contact", "RS-485", "TTL/GND"};
-int CPCOCam::OnSelectSignalType(MM::PropertyBase* pProp, MM::ActionType eAct)
-{
-  if(m_pCamera->m_iCamClass != 3)// pco.camera
-    return DEVICE_OK;
-
-  long isc = 0;//GetSignalNum(pProp->GetName());
-  string szval;
-  pProp->Get(szval);
-  MM::Property *pprophelper = (MM::Property *)pProp;
-  pprophelper->GetData(szval.c_str(), isc);
-  isc /= 0x100;
-  isc--;
-
-  if(eAct == MM::BeforeGet)
-  {
-    int iselectedpar = iBitToIndex[m_pCamera->m_strCamera.strTiming.strSignal[isc].wType];
-    pProp->Set(szSelectSignalType[iselectedpar]);
-  }
-  else if(eAct == MM::AfterSet)
-  {
-    std::string szselectedsignal;
-
-    pProp->Get(szselectedsignal);
-
-    string szsignal;
-    int inewval = 1;
-    for(int i = 0; i < 5; i++)
-    {
-      szsignal = szSelectSignalType[i];
-      if(szsignal == szselectedsignal)
-      {
-        if(m_pCamera->m_strCamera.strTiming.strSignal[isc].wType != (WORD) inewval)
-        {
-          m_pCamera->m_strCamera.strTiming.strSignal[isc].wType = (WORD) inewval;
-          SetupCamera(true, false);
-        }
-        break;
-      }
-      inewval *= 2;
-    }
-  }
-
-  return DEVICE_OK;
-}
-
-char szSelectSignalFilter[3][40] = {"off", "medium", "high"};
-int CPCOCam::OnSelectSignalFilter(MM::PropertyBase* pProp, MM::ActionType eAct)
-{
-  if(m_pCamera->m_iCamClass != 3)// pco.camera
-    return DEVICE_OK;
-
-  long isc = 0;//GetSignalNum(pProp->GetName());
-  string szval;
-  pProp->Get(szval);
-  MM::Property *pprophelper = (MM::Property *)pProp;
-  pprophelper->GetData(szval.c_str(), isc);
-  isc /= 0x100;
-  isc--;
-
-  if(eAct == MM::BeforeGet)
-  {
-    int iselectedpar = iBitToIndex[m_pCamera->m_strCamera.strTiming.strSignal[isc].wFilterSetting];
-    pProp->Set(szSelectSignalFilter[iselectedpar]);
-  }
-  else if(eAct == MM::AfterSet)
-  {
-    std::string szselectedsignal;
-
-    pProp->Get(szselectedsignal);
-
-    string szsignal;
-    int inewval = 1;
-    for(int i = 0; i < 3; i++)
-    {
-      szsignal = szSelectSignalFilter[i];
-      if(szsignal == szselectedsignal)
-      {
-        if(m_pCamera->m_strCamera.strTiming.strSignal[isc].wFilterSetting != (WORD) inewval)
-        {
-          m_pCamera->m_strCamera.strTiming.strSignal[isc].wFilterSetting = (WORD) inewval;
-          SetupCamera(true, false);
-        }
-        break;
-      }
-      inewval *= 2;
-    }
-  }
-
-  return DEVICE_OK;
-}
-
-char szSelectSignalPolarity[4][40] = {"high", "low", "rising", "falling"};
-int CPCOCam::OnSelectSignalPolarity(MM::PropertyBase* pProp, MM::ActionType eAct)
-{
-  if(m_pCamera->m_iCamClass != 3)// pco.camera
-    return DEVICE_OK;
-
-  long isc = 0;//GetSignalNum(pProp->GetName());
-  string szval;
-  pProp->Get(szval);
-  MM::Property *pprophelper = (MM::Property *)pProp;
-  pprophelper->GetData(szval.c_str(), isc);
-  isc /= 0x100;
-  isc--;
-
-  if(eAct == MM::BeforeGet)
-  {
-    int iselectedpar = iBitToIndex[m_pCamera->m_strCamera.strTiming.strSignal[isc].wPolarity];
-    pProp->Set(szSelectSignalPolarity[iselectedpar]);
-  }
-  else if(eAct == MM::AfterSet)
-  {
-    std::string szselectedsignal;
-
-    pProp->Get(szselectedsignal);
-
-    string szsignal;
-    int inewval = 1;
-    for(int i = 0; i < 4; i++)
-    {
-      szsignal = szSelectSignalPolarity[i];
-      if(szsignal == szselectedsignal)
-      {
-        if(m_pCamera->m_strCamera.strTiming.strSignal[isc].wPolarity != (WORD) inewval)
-        {
-          m_pCamera->m_strCamera.strTiming.strSignal[isc].wPolarity = (WORD) inewval;
-          SetupCamera(true, false);
-        }
-        break;
-      }
-      inewval *= 2;
-    }
-  }
-
-  return DEVICE_OK;
-}
-
-
-int CPCOCam::InitHWIO()
-{
-  CPropertyAction* pAct;
-  int nRet;
-  if(m_pCamera->m_strCamera.strSensor.strDescription.dwGeneralCapsDESC1 & GENERALCAPS1_HW_IO_SIGNAL_DESCRIPTOR)
-  {
-    for(int isc = 0; isc < m_pCamera->m_strCamera.strSensor.strSignalDesc.wNumOfSignals; isc++)
-    {
-      char csh[HWIOBUFLEN], csh2[HWIOBUFLEN];
-      int ivalue_helper;
-
-      string szSignalName, szhelp;
-
-      ivalue_helper = (isc + 1) * 0x100;
-      sprintf_s(csh, HWIOBUFLEN, "Signal %d (%s)", isc + 1, m_pCamera->m_strCamera.strSensor.strSignalDesc.strSingeSignalDesc[isc].strSignalName[0]);
-      szSignalName = csh;
-
-      pAct = new CPropertyAction(this, &CPCOCam::OnSelectSignalOnOff);
-      sprintf_s(csh, HWIOBUFLEN, "%s Status", szSignalName.c_str());
-
-      if(m_pCamera->m_strCamera.strTiming.strSignal[isc].wEnabled == 0)
-        szhelp = "off";
-      else
-        szhelp = "on";
-
-      nRet = CreateProperty(csh, szhelp.c_str(), MM::String, false, pAct);
-      if(nRet != DEVICE_OK)
-        return nRet;
-      nRet = AddAllowedValue(csh, "off", 0 + ivalue_helper);
-      if(nRet != DEVICE_OK)
-        return nRet;
-      nRet = AddAllowedValue(csh, "on", 1 + ivalue_helper);
-      if(nRet != DEVICE_OK)
-        return nRet;
-
-      if(m_pCamera->m_strCamera.strSensor.strSignalDesc.strSingeSignalDesc[isc].strSignalName[1][0] != 0) // More than one signal name
-      {
-        int i;
-        pAct = new CPropertyAction(this, &CPCOCam::OnSelectSignal);
-        sprintf_s(csh, HWIOBUFLEN, "%s Selection", szSignalName.c_str());
-        sprintf_s(csh2, HWIOBUFLEN, "%s", m_pCamera->m_strCamera.strSensor.strSignalDesc.strSingeSignalDesc[isc].strSignalName[m_pCamera->m_strCamera.strTiming.strSignal[isc].wSelected]);
-        nRet = CreateProperty(csh, csh2, MM::String, false, pAct);
-        if(nRet != DEVICE_OK)
-          return nRet;
-
-        for(i = 0; i < 4; i++)
-        {
-          if(m_pCamera->m_strCamera.strSensor.strSignalDesc.strSingeSignalDesc[isc].strSignalName[i][0] != 0)
-          {
-            sprintf_s(csh2, HWIOBUFLEN, "%s", m_pCamera->m_strCamera.strSensor.strSignalDesc.strSingeSignalDesc[isc].strSignalName[i]);
-
-            nRet = AddAllowedValue(csh, csh2, i + ivalue_helper);
-            if(nRet != DEVICE_OK)
-              return nRet;
-          }
-          else
-            break;
-        }
-        if(m_pCamera->m_strCamera.strTiming.strSignal[isc].wSelected >= i)
-          m_pCamera->m_strCamera.strTiming.strSignal[isc].wSelected = 0;
-      }
-      else
-        m_pCamera->m_strCamera.strTiming.strSignal[isc].wSelected = 0;
-
-      int isignal = m_pCamera->m_strCamera.strTiming.strSignal[isc].wSelected;
-      int iflag = 1 << isignal;
-      if(iflag & m_pCamera->m_strCamera.strSensor.strSignalDesc.strSingeSignalDesc[isc].wSignalDefinitions)
-      {
-        if((m_pCamera->m_strCamera.strTiming.strSignal[isc].dwSignalFunctionality[0] == 0x07) ||
-          (m_pCamera->m_strCamera.strTiming.strSignal[isc].dwSignalFunctionality[1] == 0x07) ||
-          (m_pCamera->m_strCamera.strTiming.strSignal[isc].dwSignalFunctionality[2] == 0x07) ||
-          (m_pCamera->m_strCamera.strTiming.strSignal[isc].dwSignalFunctionality[3] == 0x07))
-        {
-          int isignaltextindex = m_pCamera->m_strCamera.strTiming.strSignal[isc].dwParameter[isignal] - 1;
-          sprintf_s(csh, HWIOBUFLEN, "%s Timing", szSignalName.c_str());
-
-          pAct = new CPropertyAction(this, &CPCOCam::OnSelectSignalTiming);
-          if((isignaltextindex >= 0) && (isignaltextindex <= 3))
-            nRet = CreateProperty(csh, szSelectSignalTiming[isignaltextindex], MM::String, false, pAct);
-          else
-            nRet = CreateProperty(csh, "Not available", MM::String, false, pAct);
-          if(nRet != DEVICE_OK)
-            return nRet;
-
-          if(m_pCamera->m_strCamera.strTiming.strSignal[isc].dwSignalFunctionality[isignal] == 0x07)
-          {
-            // We have got timing output settings
-
-            nRet = AddAllowedValue(csh, szSelectSignalTiming[0], 0 + ivalue_helper);
-            if(nRet != DEVICE_OK)
-              return nRet;
-            nRet = AddAllowedValue(csh, szSelectSignalTiming[1], 1 + ivalue_helper);
-            if(nRet != DEVICE_OK)
-              return nRet;
-            nRet = AddAllowedValue(csh, szSelectSignalTiming[2], 2 + ivalue_helper);
-            if(nRet != DEVICE_OK)
-              return nRet;
-            nRet = AddAllowedValue(csh, szSelectSignalTiming[3], 3 + ivalue_helper);
-            if(nRet != DEVICE_OK)
-              return nRet;
-          }
-          else
-          {
-            nRet = AddAllowedValue(csh, "Not available", ivalue_helper);
-            if(nRet != DEVICE_OK)
-              return nRet;
-          }
-        }
-      }
-
-      int count = 0;
-      WORD wsignal = m_pCamera->m_strCamera.strSensor.strSignalDesc.strSingeSignalDesc[isc].wSignalTypes;
-      WORD wh = wsignal;
-      wh = wh - ((wh >> 1) & 0x5555);    // 0101 0101 0101 0101 
-      wh = (wh & 0x3333) + ((wh >> 2) & 0x3333);// 0011 0011 0011 0011
-      wh = (wh + (wh >> 4)) & 0x0F0F;  // 0000 1111 0000 1111
-      count = (wh + (wh >> 8)) & 0xFF;
-
-      // If wh has got more than 1 bit set, we can choose. Also the current signal type bit must be in descriptors signal types.
-      if((count > 1) && (m_pCamera->m_strCamera.strTiming.strSignal[isc].wType & wsignal))
-      {
-        pAct = new CPropertyAction(this, &CPCOCam::OnSelectSignalType);
-        sprintf_s(csh, HWIOBUFLEN, "%s Type", szSignalName.c_str());
-
-        nRet = CreateProperty(csh, szSelectSignalType[iBitToIndex[m_pCamera->m_strCamera.strTiming.strSignal[isc].wType]], MM::String, false, pAct);
-        if(nRet != DEVICE_OK)
-          return nRet;
-        if(m_pCamera->m_strCamera.strSensor.strSignalDesc.strSingeSignalDesc[isc].wSignalTypes & SIGNAL_TYPE_TTL)
-        {
-          nRet = AddAllowedValue(csh, szSelectSignalType[0], 0 + ivalue_helper);
-          if(nRet != DEVICE_OK)
-            return nRet;
-        }
-        if(m_pCamera->m_strCamera.strSensor.strSignalDesc.strSingeSignalDesc[isc].wSignalTypes & SIGNAL_TYPE_HL_SIG)
-        {
-          nRet = AddAllowedValue(csh, szSelectSignalType[1], 1 + ivalue_helper);
-          if(nRet != DEVICE_OK)
-            return nRet;
-        }
-        if(m_pCamera->m_strCamera.strSensor.strSignalDesc.strSingeSignalDesc[isc].wSignalTypes & SIGNAL_TYPE_CONTACT)
-        {
-          nRet = AddAllowedValue(csh, szSelectSignalType[2], 2 + ivalue_helper);
-          if(nRet != DEVICE_OK)
-            return nRet;
-        }
-        if(m_pCamera->m_strCamera.strSensor.strSignalDesc.strSingeSignalDesc[isc].wSignalTypes & SIGNAL_TYPE_RS485)
-        {
-          nRet = AddAllowedValue(csh, szSelectSignalType[3], 3 + ivalue_helper);
-          if(nRet != DEVICE_OK)
-            return nRet;
-        }
-        if(m_pCamera->m_strCamera.strSensor.strSignalDesc.strSingeSignalDesc[isc].wSignalTypes & SIGNAL_TYPE_TTL_A_GND_B)
-        {
-          nRet = AddAllowedValue(csh, szSelectSignalType[4], 4 + ivalue_helper);
-          if(nRet != DEVICE_OK)
-            return nRet;
-        }
-      }
-      WORD wfilter = m_pCamera->m_strCamera.strSensor.strSignalDesc.strSingeSignalDesc[isc].wSignalFilter;
-      wh = wfilter;
-      wh = wh - ((wh >> 1) & 0x5555);    // 0101 0101 0101 0101 
-      wh = (wh & 0x3333) + ((wh >> 2) & 0x3333);// 0011 0011 0011 0011
-      wh = (wh + (wh >> 4)) & 0x0F0F;  // 0000 1111 0000 1111
-      count = (wh + (wh >> 8)) & 0xFF;
-
-      // If wh has got more than 1 bit set, we can choose. Also the current signal filter bit must be in descriptors signal filters.
-      if((count > 1) && (m_pCamera->m_strCamera.strTiming.strSignal[isc].wFilterSetting & wfilter))
-      {
-        pAct = new CPropertyAction(this, &CPCOCam::OnSelectSignalFilter);
-        sprintf_s(csh, HWIOBUFLEN, "%s Filter", szSignalName.c_str());
-
-        nRet = CreateProperty(csh, szSelectSignalFilter[iBitToIndex[m_pCamera->m_strCamera.strTiming.strSignal[isc].wFilterSetting]], MM::String, false, pAct);
-        if(nRet != DEVICE_OK)
-          return nRet;
-        if(m_pCamera->m_strCamera.strSensor.strSignalDesc.strSingeSignalDesc[isc].wSignalFilter & SIGNAL_FILTER_OFF)
-        {
-          nRet = AddAllowedValue(csh, szSelectSignalFilter[0], 0 + ivalue_helper);
-          if(nRet != DEVICE_OK)
-            return nRet;
-        }
-        if(m_pCamera->m_strCamera.strSensor.strSignalDesc.strSingeSignalDesc[isc].wSignalFilter & SIGNAL_FILTER_MED)
-        {
-          nRet = AddAllowedValue(csh, szSelectSignalFilter[1], 1 + ivalue_helper);
-          if(nRet != DEVICE_OK)
-            return nRet;
-        }
-        if(m_pCamera->m_strCamera.strSensor.strSignalDesc.strSingeSignalDesc[isc].wSignalFilter & SIGNAL_FILTER_HIGH)
-        {
-          nRet = AddAllowedValue(csh, szSelectSignalFilter[2], 2 + ivalue_helper);
-          if(nRet != DEVICE_OK)
-            return nRet;
-        }
-      }
-
-      WORD wpolarity = m_pCamera->m_strCamera.strSensor.strSignalDesc.strSingeSignalDesc[isc].wSignalPolarity;
-      wh = wpolarity;
-      wh = wh - ((wh >> 1) & 0x5555);    // 0101 0101 0101 0101 
-      wh = (wh & 0x3333) + ((wh >> 2) & 0x3333);// 0011 0011 0011 0011
-      wh = (wh + (wh >> 4)) & 0x0F0F;  // 0000 1111 0000 1111
-      count = (wh + (wh >> 8)) & 0xFF;
-
-      // If wh has got more than 1 bit set, we can choose. Also the current signal polarity bit must be in descriptors signal polarities.
-      if((count > 1) && (m_pCamera->m_strCamera.strTiming.strSignal[isc].wPolarity & wpolarity))
-      {
-        pAct = new CPropertyAction(this, &CPCOCam::OnSelectSignalPolarity);
-        sprintf_s(csh, HWIOBUFLEN, "%s Polarity", szSignalName.c_str());
-
-        nRet = CreateProperty(csh, szSelectSignalPolarity[iBitToIndex[m_pCamera->m_strCamera.strTiming.strSignal[isc].wPolarity]], MM::String, false, pAct);
-        if(nRet != DEVICE_OK)
-          return nRet;
-        if(m_pCamera->m_strCamera.strSensor.strSignalDesc.strSingeSignalDesc[isc].wSignalPolarity & SIGNAL_POL_HIGH)
-        {
-          nRet = AddAllowedValue(csh, szSelectSignalPolarity[0], 0 + ivalue_helper);
-          if(nRet != DEVICE_OK)
-            return nRet;
-        }
-        if(m_pCamera->m_strCamera.strSensor.strSignalDesc.strSingeSignalDesc[isc].wSignalPolarity & SIGNAL_POL_LOW)
-        {
-          nRet = AddAllowedValue(csh, szSelectSignalPolarity[1], 1 + ivalue_helper);
-          if(nRet != DEVICE_OK)
-            return nRet;
-        }
-        if(m_pCamera->m_strCamera.strSensor.strSignalDesc.strSingeSignalDesc[isc].wSignalPolarity & SIGNAL_POL_RISE)
-        {
-          nRet = AddAllowedValue(csh, szSelectSignalPolarity[2], 2 + ivalue_helper);
-          if(nRet != DEVICE_OK)
-            return nRet;
-        }
-        if(m_pCamera->m_strCamera.strSensor.strSignalDesc.strSingeSignalDesc[isc].wSignalPolarity & SIGNAL_POL_FALL)
-        {
-          nRet = AddAllowedValue(csh, szSelectSignalPolarity[3], 3 + ivalue_helper);
-          if(nRet != DEVICE_OK)
-            return nRet;
-        }
-      }
-    }
-  }
-
-  return DEVICE_OK;
-}
-
-int CPCOCam::InitLineTiming()
-{
-  CPropertyAction* pAct;
-  int nRet;
-  string szhelp;
-  string csh;
-  char szh[100];
-
-  if((m_pCamera->m_strCamera.strAPIManager.wAPIManagementFlags & APIMANAGEMENTFLAG_LINE_TIMING) == APIMANAGEMENTFLAG_LINE_TIMING)
-  {
-    pAct = new CPropertyAction(this, &CPCOCam::OnCmosParameter);
-
-    if(m_pCamera->m_strCamera.strTiming.wCMOSParameter == 0)
-      szhelp = "Off";
-    else
-      szhelp = "On";
-    csh = "Light Sheet Mode";
-    nRet = CreateProperty(csh.c_str(), szhelp.c_str(), MM::String, false, pAct);
-    if(nRet != DEVICE_OK)
-      return nRet;
-    nRet = AddAllowedValue(csh.c_str(), "Off", 0);
-    if(nRet != DEVICE_OK)
-      return nRet;
-    nRet = AddAllowedValue(csh.c_str(), "On", 1);
-    if(nRet != DEVICE_OK)
-      return nRet;
-
-    pAct = new CPropertyAction(this, &CPCOCam::OnCmosTimeBase);
-
-    if(m_pCamera->m_strCamera.strTiming.wCMOSTimeBase == 0)
-      szhelp = "ns";
-    else
-      if(m_pCamera->m_strCamera.strTiming.wCMOSTimeBase == 1)
-        szhelp = "us";
-      else
-        szhelp = "ms";
-    csh = "Light Sheet Mode Timebase";
-    nRet = CreateProperty(csh.c_str(), szhelp.c_str(), MM::String, false, pAct);
-    if(nRet != DEVICE_OK)
-      return nRet;
-    nRet = AddAllowedValue(csh.c_str(), "ns", 0);
-    if(nRet != DEVICE_OK)
-      return nRet;
-    nRet = AddAllowedValue(csh.c_str(), "us", 1);
-    if(nRet != DEVICE_OK)
-      return nRet;
-    nRet = AddAllowedValue(csh.c_str(), "ms", 2);
-    if(nRet != DEVICE_OK)
-      return nRet;
-
-
-    pAct = new CPropertyAction(this, &CPCOCam::OnCmosLineTime);
-    csh = "Light Sheet Mode Line Time";
-    sprintf_s(szh, 100, "%d", m_dwCMOSLineTime);
-    nRet = CreateProperty(csh.c_str(), szh, MM::Integer, false, pAct);
-    if(nRet != DEVICE_OK)
-      return nRet;
-
-    pAct = new CPropertyAction(this, &CPCOCam::OnCmosDelayLines);
-    csh = "Light Sheet Mode Delay Lines";
-    sprintf_s(szh, 100, "%d", m_dwCMOSDelayLines);
-    nRet = CreateProperty(csh.c_str(), szh, MM::Integer, false, pAct);
-    if(nRet != DEVICE_OK)
-      return nRet;
-
-    pAct = new CPropertyAction(this, &CPCOCam::OnCmosExposureLines);
-    csh = "Light Sheet Mode Exposure Lines";
-    sprintf_s(szh, 100, "%d", m_dwCMOSExposureLines);
-    nRet = CreateProperty(csh.c_str(), szh, MM::Integer, false, pAct);
-    if(nRet != DEVICE_OK)
-      return nRet;
-  }
-
-  return DEVICE_OK;
-}
-
-int CPCOCam::InitFlim()
-{
-  //bool bshutterchanged = false;
-  //bool breadoutchanged = false;
-  WORD wType = 0, wLen = 20;
-  DWORD dwSetup[20] = {0, 0, 0, 0, 0,
-    0, 0, 0, 0, 0,
-    0, 0, 0, 0, 0,
-    0, 0, 0, 0, 0};
-  int err = PCO_NOERROR;
-
-  err = m_pCamera->GetCameraSetup(&wType, &dwSetup[0], &wLen);
-  m_wFlimSourceSelect = (WORD) dwSetup[0]; // FLIM_MODULATION_SOURCE_EXTERN: 1; FLIM_MODULATION_SOURCE_INTERN: 0;
-  m_wFlimOutputWaveform = (WORD) dwSetup[1]; // FLIM_MODULATION_OUTPUT_WAVEFORM_NONE: 0; FLIM_MODULATION_OUTPUT_WAVEFORM_SINE: 1; FLIM_MODULATION_OUTPUT_WAVEFORM_RECT: 2
-  m_wFlimPhaseNumber = (WORD) dwSetup[2]; // FLIM_PHASE_MANUAL_SHIFTING: 0; FLIM_PHASE_NUMBER_2: 1; FLIM_PHASE_NUMBER_4: 2; FLIM_PHASE_NUMBER_8: 3; FLIM_PHASE_NUMBER_16: 4
-
-  m_wFlimPhaseSymmetry = (WORD) dwSetup[3]; // FLIM_PHASE_SYMMETRY_SINGULAR: 0; FLIM_PHASE_SYMMETRY_TWICE: 1;
-
-  m_wFlimPhaseOrder = (WORD) dwSetup[4]; // FLIM_PHASE_ORDER_ASCENDING: 0; FLIM_PHASE_ORDER_OPPOSITE:1;
-
-  m_wFlimTapSelect = (WORD) dwSetup[5]; // FLIM_TAP_SELECT_BOTH: 0; FLIM_TAP_SELECT_0: 1; FLIM_TAP_SELECT_180: 2;
-
-  m_wFlimAsymmetryCorrection = (WORD) dwSetup[6]; // FLIM_ASYMMETRY_CORRECTION_OFF: 0; FLIM_ASYMMETRY_CORRECTION_AVERAGE: 1;
-
-  m_wFlimCalculationMode = (WORD) dwSetup[7];
-  m_wFlimReferencingMode = (WORD) dwSetup[8];
-  m_wFlimThresholdLow = (WORD) dwSetup[9];
-  m_wFlimThresholdHigh = (WORD) dwSetup[10];
-  m_wFlimOutputMode = (WORD) dwSetup[11];
-
-  m_dwFlimFrequency = dwSetup[12];
-  m_dwFlimPhaseMilliDeg = dwSetup[13];
-
-  m_bFlimMasterFrequencyMHz = dwSetup[14] == 1 ? true : false;
-
-
-
-  CPropertyAction* pAct;
-  int nRet;
-  string szhelp;
-  string csh;
-  char szh[100];
-
-  pAct = new CPropertyAction(this, &CPCOCam::OnFlimModulationSource);
-
-  if(m_wFlimSourceSelect == 0)
-    szhelp = "(0) intern";
-  else
-    szhelp = "(1) extern";
-  csh = "Flim (01) Modulation Source";
-  nRet = CreateProperty(csh.c_str(), szhelp.c_str(), MM::String, false, pAct);
-  if(nRet != DEVICE_OK)
-    return nRet;
-  nRet = AddAllowedValue(csh.c_str(), "(0) intern", 0);
-  if(nRet != DEVICE_OK)
-    return nRet;
-  nRet = AddAllowedValue(csh.c_str(), "(1) extern", 1);
-  if(nRet != DEVICE_OK)
-    return nRet;
-
-  pAct = new CPropertyAction(this, &CPCOCam::OnFlimFrequency);
-  csh = "Flim (02) Master Frequency";
-  sprintf_s(szh, 100, "%d", m_dwFlimFrequency);
-  nRet = CreateProperty(csh.c_str(), szh, MM::Integer, false, pAct);
-  if(nRet != DEVICE_OK)
-    return nRet;
-
-
-  pAct = new CPropertyAction(this, &CPCOCam::OnFlimMasterFrequencyMHz);
-  CreateProperty("Flim (03) Master Frequency [MHz/kHz]", "(0) kHz", MM::String, false, pAct);
-  AddAllowedValue("Flim (03) Master Frequency [MHz/kHz]", "(0) kHz", 0);
-  AddAllowedValue("Flim (03) Master Frequency [MHz/kHz]", "(1) MHz", 1);
-
-  pAct = new CPropertyAction(this, &CPCOCam::OnFlimRelativePhase);
-  csh = "Flim (04) Relative Phase";
-  sprintf_s(szh, 100, "%d", m_dwFlimPhaseMilliDeg);
-  nRet = CreateProperty(csh.c_str(), szh, MM::Integer, false, pAct);
-  if(nRet != DEVICE_OK)
-    return nRet;
-
-  pAct = new CPropertyAction(this, &CPCOCam::OnFlimOutputWaveForm);
-  if(m_wFlimOutputWaveform == 0)
-    szhelp = "(0) none";
-  else
-    if(m_wFlimOutputWaveform == 1)
-      szhelp = "(1) sine wave";
-    else
-      szhelp = "(2) square wave";
-  csh = "Flim (05) Output Waveform";
-  nRet = CreateProperty(csh.c_str(), szhelp.c_str(), MM::String, false, pAct);
-  if(nRet != DEVICE_OK)
-    return nRet;
-  nRet = AddAllowedValue(csh.c_str(), "(0) none", 0);
-  if(nRet != DEVICE_OK)
-    return nRet;
-  nRet = AddAllowedValue(csh.c_str(), "(1) sine wave", 1);
-  if(nRet != DEVICE_OK)
-    return nRet;
-  nRet = AddAllowedValue(csh.c_str(), "(2) square wave", 2);
-
-  switch(m_wFlimPhaseNumber)
-  {
-    case 0:
-    szhelp = "(0) shiftable pair";
-    break;
-    case 1:
-    szhelp = "(1) 2";
-    break;
-    case 2:
-    szhelp = "(2) 4";
-    break;
-    case 3:
-    szhelp = "(3) 8";
-    break;
-    case 4:
-    szhelp = "(4) 16";
-    break;
-  }
-  csh = "Flim (06) Number of Phase Samples";
-  pAct = new CPropertyAction(this, &CPCOCam::OnFlimNumberOfPhaseSamples);
-  nRet = CreateProperty(csh.c_str(), szhelp.c_str(), MM::String, false, pAct);
-  if(nRet != DEVICE_OK)
-    return nRet;
-  nRet = AddAllowedValue(csh.c_str(), "(0) shiftable pair", 0);
-  if(nRet != DEVICE_OK)
-    return nRet;
-  nRet = AddAllowedValue(csh.c_str(), "(1) 2", 1);
-  if(nRet != DEVICE_OK)
-    return nRet;
-  nRet = AddAllowedValue(csh.c_str(), "(2) 4", 2);
-  if(nRet != DEVICE_OK)
-    return nRet;
-  nRet = AddAllowedValue(csh.c_str(), "(3) 8", 3);
-  if(nRet != DEVICE_OK)
-    return nRet;
-  nRet = AddAllowedValue(csh.c_str(), "(4) 16", 4);
-  if(nRet != DEVICE_OK)
-    return nRet;
-
-  if(m_wFlimPhaseSymmetry == 0)
-    szhelp = "(0) no";
-  else
-    szhelp = "(1) yes";
-  csh = "Flim (07) Additional Phase Sampling";
-  pAct = new CPropertyAction(this, &CPCOCam::OnFlimPhaseSymmetry);
-  nRet = CreateProperty(csh.c_str(), szhelp.c_str(), MM::String, false, pAct);
-  if(nRet != DEVICE_OK)
-    return nRet;
-  nRet = AddAllowedValue(csh.c_str(), "(0) no", 0);
-  if(nRet != DEVICE_OK)
-    return nRet;
-  nRet = AddAllowedValue(csh.c_str(), "(1) yes", 1);
-  if(nRet != DEVICE_OK)
-    return nRet;
-
-  if(m_wFlimPhaseOrder == 0)
-    szhelp = "(0) ascending";
-  else
-    szhelp = "(1) opposite";
-  csh = "Flim (08) Phase Order";
-  pAct = new CPropertyAction(this, &CPCOCam::OnFlimPhaseOrder);
-  nRet = CreateProperty(csh.c_str(), szhelp.c_str(), MM::String, false, pAct);
-  if(nRet != DEVICE_OK)
-    return nRet;
-  nRet = AddAllowedValue(csh.c_str(), "(0) ascending", 0);
-  if(nRet != DEVICE_OK)
-    return nRet;
-  nRet = AddAllowedValue(csh.c_str(), "(1) opposite", 1);
-  if(nRet != DEVICE_OK)
-    return nRet;
-
-  switch(m_wFlimTapSelect)
-  {
-    case 0:
-    szhelp = "(0) Tap A + B";
-    break;
-    case 1:
-    szhelp = "(1) Tap A";
-    break;
-    case 2:
-    szhelp = "(2) Tap B";
-    break;
-  }
-  csh = "Flim (09) Tap Selection";
-  pAct = new CPropertyAction(this, &CPCOCam::OnFlimTapSelection);
-  nRet = CreateProperty(csh.c_str(), szhelp.c_str(), MM::String, false, pAct);
-  if(nRet != DEVICE_OK)
-    return nRet;
-  nRet = AddAllowedValue(csh.c_str(), "(0) Tap A + B", 0);
-  if(nRet != DEVICE_OK)
-    return nRet;
-  nRet = AddAllowedValue(csh.c_str(), "(1) Tap A", 1);
-  if(nRet != DEVICE_OK)
-    return nRet;
-  nRet = AddAllowedValue(csh.c_str(), "(2) Tap B", 2);
-  if(nRet != DEVICE_OK)
-    return nRet;
-
-  if(m_wFlimAsymmetryCorrection == 0)
-    szhelp = "(0) off";
-  else
-    szhelp = "(1) on";
-  csh = "Flim (10) Asym. Correction in Camera";
-  pAct = new CPropertyAction(this, &CPCOCam::OnFlimAsymCorrection);
-  nRet = CreateProperty(csh.c_str(), szhelp.c_str(), MM::String, false, pAct);
-  if(nRet != DEVICE_OK)
-    return nRet;
-  nRet = AddAllowedValue(csh.c_str(), "(0) off", 0);
-  if(nRet != DEVICE_OK)
-    return nRet;
-  nRet = AddAllowedValue(csh.c_str(), "(1) on", 1);
-  return nRet;
-}
-
-int CPCOCam::SetupFlim()
-{
-  //bool bshutterchanged = false;
-  //bool breadoutchanged = false;
-  WORD wType = 0, wLen = 20;
-  DWORD dwSetup[20] = {0, 0, 0, 0, 0,
-    0, 0, 0, 0, 0,
-    0, 0, 0, 0, 0,
-    0, 0, 0, 0, 0};
-  int err = PCO_NOERROR;
-  DWORD dwflags = 0;
-
-  if(m_wFlimPhaseNumber == 0)      // Must be 2,4,8 or 16 (!=0)
-  {
-    m_wFlimPhaseSymmetry = 0;
-    m_wFlimPhaseOrder = 0;
-    m_wFlimAsymmetryCorrection = 0;
-  }
-
-  if(m_wFlimPhaseSymmetry == 0)
-  {
-    m_wFlimAsymmetryCorrection = 0;
-    m_wFlimPhaseOrder = 0;
-  }
-
-  if(m_wFlimPhaseOrder != 1)         // Must be Opposite (1)
-    m_wFlimAsymmetryCorrection = 0;
-
-  if(m_wFlimTapSelect != 0)          // Must be Tap A + B (0)
-    m_wFlimAsymmetryCorrection = 0;
-
-  err = m_pCamera->GetCameraSetup(&wType, &dwSetup[0], &wLen);
-  dwSetup[0] = m_wFlimSourceSelect;
-  dwSetup[1] = m_wFlimOutputWaveform;
-  dwSetup[2] = m_wFlimPhaseNumber;
-  dwSetup[3] = m_wFlimPhaseSymmetry;
-  dwSetup[4] = m_wFlimPhaseOrder;
-  dwSetup[5] = m_wFlimTapSelect;
-  dwSetup[6] = m_wFlimAsymmetryCorrection;
-  dwSetup[7] = m_wFlimCalculationMode;
-  dwSetup[8] = m_wFlimReferencingMode;
-  dwSetup[9] = m_wFlimThresholdLow;
-  dwSetup[10] = m_wFlimThresholdHigh;
-  dwSetup[11] = m_wFlimOutputMode;
-  dwSetup[12] = m_dwFlimFrequency;
-  dwSetup[13] = m_dwFlimPhaseMilliDeg;
-  dwSetup[14] = m_bFlimMasterFrequencyMHz ? 1 : 0;
-  err = m_pCamera->SetCameraSetup(wType, &dwSetup[0], wLen, dwflags);
-  UpdateStatus();
-  return 0;
-}
-
-int CPCOCam::SetupFanLED()
-{
-  DWORD dwLedParam, dwflagsout = 0;
-  WORD wMode, wValue, wmin, wmax;
-  wMode = m_wFanControl;
-  wValue = m_wFanSpeed;
-  dwLedParam = m_dwLEDControl;
-  return m_pCamera->HandleFanAndLED(2, &wMode, &wValue, &wmin, &wmax, &dwLedParam, &dwflagsout);
-}
-
-int CPCOCam::OnFanControl(MM::PropertyBase* pProp, MM::ActionType eAct)
-{
-  if (eAct == MM::BeforeGet)
-  {
-    if (m_wFanControl == FAN_CONTROL_MODE_AUTO)
-      pProp->Set("On");
-    else
-      pProp->Set("Off");
-  }
-  else if (eAct == MM::AfterSet)
-  {
-    string tmp;
-    long ModeTmp;
-    int ihelp;
-    pProp->Get(tmp);
-    ((MM::Property*) pProp)->GetData(tmp.c_str(), ModeTmp);
-    ihelp = (ModeTmp == 1);
-
-    if (ihelp != 0)
-      m_wFanControl = FAN_CONTROL_MODE_AUTO;
-    else
-      m_wFanControl = FAN_CONTROL_MODE_USER;
-
-    int nErr = SetupFanLED();
-
-    if (nErr != 0)
-      return nErr;
-  }
-  return DEVICE_OK;
-}
-
-int CPCOCam::OnFanSpeed(MM::PropertyBase* pProp, MM::ActionType eAct)
-{
-  if (eAct == MM::BeforeGet)
-  {
-    pProp->Set((long)m_wFanSpeed);
-  }
-  else if (eAct == MM::AfterSet)
-  {
-    int nErr = 0;
-    long lhelp;
-    WORD whelp;
-    pProp->Get(lhelp);
-    whelp = (WORD)lhelp;
-    if (lhelp != m_wFanSpeed)
-    {
-      m_wFanSpeed = whelp;
-      nErr = SetupFanLED();
-    }
-
-    if (nErr != 0)
-      return nErr;
-  }
-  return DEVICE_OK;
-}
-
-int CPCOCam::OnLEDControl(MM::PropertyBase* pProp, MM::ActionType eAct)
-{
-  if (eAct == MM::BeforeGet)
-  {
-    if (m_dwLEDControl == HW_LED_SIGNAL_ON)
-      pProp->Set("On");
-    else
-      pProp->Set("Off");
-  }
-  else if (eAct == MM::AfterSet)
-  {
-    string tmp;
-    long ModeTmp;
-    int ihelp;
-    pProp->Get(tmp);
-    ((MM::Property*) pProp)->GetData(tmp.c_str(), ModeTmp);
-    ihelp = (ModeTmp == 1);
-
-    if (ihelp != 0)
-      m_dwLEDControl = HW_LED_SIGNAL_ON;
-    else
-      m_dwLEDControl = HW_LED_SIGNAL_OFF;
-
-    int nErr = SetupFanLED();
-
-    if (nErr != 0)
-      return nErr;
-  }
-  return DEVICE_OK;
-}
-
-int CPCOCam::InitFanLED()
-{
-  DWORD dwLedParam, dwflagsout = 0;
-  WORD wMode, wValue, wmin, wmax;
-  m_pCamera->HandleFanAndLED(1, &wMode, &wValue, &wmin, &wmax, &dwLedParam, &dwflagsout);
-  if ((dwflagsout & FAN_IS_POSSIBLE) != FAN_IS_POSSIBLE)
-    return 0;
-
-  CPropertyAction* pAct;
-  int nRet;
-  string szhelp;
-  string csh;
-  char szh[100];
-
-  m_wFanControl = wMode;
-  m_wFanSpeed = wValue;
-  m_dwLEDControl = dwLedParam;
-
-  pAct = new CPropertyAction(this, &CPCOCam::OnFanControl);
-
-  if (wMode == FAN_CONTROL_MODE_AUTO)
-    szhelp = "On";
-  else
-    szhelp = "Off";
-  csh = "Fan/LED: Automatic Fan Control";
-  nRet = CreateProperty(csh.c_str(), szhelp.c_str(), MM::String, false, pAct);
-  if (nRet != DEVICE_OK)
-    return nRet;
-  nRet = AddAllowedValue(csh.c_str(), "Off", 0);
-  if (nRet != DEVICE_OK)
-    return nRet;
-  nRet = AddAllowedValue(csh.c_str(), "On", 1);
-  if (nRet != DEVICE_OK)
-    return nRet;
-
-  pAct = new CPropertyAction(this, &CPCOCam::OnFanSpeed);
-
-  sprintf_s(szh, 100, "%d", wValue);
-    csh = "Fan/LED: Fan Speed [%]";
-  nRet = CreateProperty(csh.c_str(), szhelp.c_str(), MM::Integer, false, pAct);
-  if (nRet != DEVICE_OK)
-    return nRet;
-  SetPropertyLimits(csh.c_str(), 0.0, 100.0);
-
-  if ((dwflagsout & LED_IS_POSSIBLE) == LED_IS_POSSIBLE)
-  {
-
-    pAct = new CPropertyAction(this, &CPCOCam::OnLEDControl);
-
-    if (dwLedParam == HW_LED_SIGNAL_ON)
-      szhelp = "On";
-    else
-      szhelp = "Off";
-    csh = "Fan/LED: Camera LED";
-    nRet = CreateProperty(csh.c_str(), szhelp.c_str(), MM::String, false, pAct);
-    if (nRet != DEVICE_OK)
-      return nRet;
-    nRet = AddAllowedValue(csh.c_str(), "Off", 0);
-    if (nRet != DEVICE_OK)
-      return nRet;
-    nRet = AddAllowedValue(csh.c_str(), "On", 1);
-    if (nRet != DEVICE_OK)
-      return nRet;
-  }
-  return DEVICE_OK;
-}
 
 ///////////////////////////////////////////////////////////////////////////////
 // Function name   : CPCOCam::Shutdown
@@ -3498,8 +1621,6 @@ const unsigned char* CPCOCam::GetImageBuffer()
   return (unsigned char*) GetBuffer(m_iLastBufferUsed[0]);
 }
 
-
-
 ///////////////////////////////////////////////////////////////////////////////
 // Function name   : char* CPCOCam::GetImageBuffer
 // Description     : Returns the raw image buffer
@@ -3516,6 +1637,8 @@ const unsigned char* CPCOCam::GetBuffer(int ibufnum)
 
   if(img_.Depth() == 2)
   {
+
+
     return (const unsigned char*) m_pic;
     //memcpy((void*) const_cast<unsigned char*>(img_.GetPixels()), (const void*)m_pic, img_.Width() * img_.Height() * 2);
   }
@@ -4125,9 +2248,29 @@ int CPCOCam::SnapImage()
   if(m_bSoftwareTriggered)
     m_pCamera->ForceTrigger();
 
-  m_iLastBufferUsed[0] = m_iNextBuffer;
+  if (m_nTrig > 1)
+  {
+    bool bok = false;
+    do
+    {
+      m_iLastBufferUsed[0] = m_iNextBuffer;
 
-  nErr = m_pCamera->WaitForImage(&m_iNextBufferToUse[0], &m_iLastBufferUsed[0]);
+      nErr = m_pCamera->WaitForImage(&m_iNextBufferToUse[0], &m_iLastBufferUsed[0]);
+      if (nErr == PCO_NOERROR)
+        bok = true;
+      if (::GetAsyncKeyState(VK_ESCAPE) & 0x8000) // Stops 'Auto Save' in progress
+      {
+        bok = true;
+        nErr = PCO_NOERROR;
+      }
+    } while (!bok);
+  }
+  else
+  {
+    m_iLastBufferUsed[0] = m_iNextBuffer;
+
+    nErr = m_pCamera->WaitForImage(&m_iNextBufferToUse[0], &m_iLastBufferUsed[0]);
+  }
   if(m_bSequenceRunning == FALSE)
   {
     if(m_bSoftwareTriggered == FALSE)
@@ -4138,6 +2281,12 @@ int CPCOCam::SnapImage()
         nErr = ilerr;
       m_bRecording = false;
     }
+  }
+  if (m_bDoDotPhoton)
+  {
+    HANDLE hcamera = m_pCamera->GetCameraHandle();
+    WORD* pImage = (WORD*)GetBuffer(m_iLastBufferUsed[0]);
+    CalcImageDotphoton((WORD*)pImage, m_iWidth * m_iHeight, hcamera);
   }
 
   if(nErr != 0)
