+++ DeviceAdapters/PCO_Generic/VersionNo.h
@@ -2,11 +2,11 @@
 #define AUTOBUILD_H
 // change the FALSE to TRUE for autoincrement of build number
 #define INCREMENT_VERSION TRUE
-#define FILEVER        2,0,0,59
-#define PRODUCTVER     2,0,0,59
-#define STRFILEVER     "2,0,0,59\0"
-#define STRPRODUCTVER  "2,0,0,59\0"
-#define FILEVER_MAJOR  2
+#define FILEVER        3,0,0,146
+#define PRODUCTVER     3,0,0,146
+#define STRFILEVER     "3,0,0,146\0"
+#define STRPRODUCTVER  "3,0,0,146\0"
+#define FILEVER_MAJOR  3
 #define FILEVER_MINOR  0
 
 #endif // AUTOBUILD_H
